Fixed Decimal Showing.

+++ StaffWorkloadSystem/StaffWorkloadSystem/ModuleStaff.aspx.cs
@@ -132,7 +132,7 @@ namespace StaffWorkloadSystem
                 {
                     //Now get all the data we can from that row.
                     int StaffID = Convert.ToInt32(dr["StaffID"]);
-                    int Weighting = Convert.ToInt32(dr["Weighting"]);
+                    decimal Weighting = Convert.ToDecimal(dr["Weighting"]);
                     bool Coordinator = dr["Weighting"].ToString() == "True" ? true : false;
                     int ExtraHours = Convert.ToInt32(dr["ExtraHours"]);
                     //Now we have the code we are missing one thing
